move send_to_segment
Issue: AAP-67366
