+++ apps/tasks/utils.py
@@ -366,75 +366,6 @@ def generate_salt() -> str:
     return str(uuid.uuid4())
 
 
-# =============================================================================
-# Segment.com Integration
-# =============================================================================
-
-
-def send_to_segment(user_id: str, event_name: str, segment_data: dict) -> str:
-    """
-    Send data to Segment.com using metrics-utility StorageSegment.
-
-    Args:
-        user_id: User ID for Segment tracking
-        event_name: Event name for tracking
-        segment_data: Dictionary of data to send
-
-    Returns:
-        str: "success" if sent, "segment_not_available" if Segment is not configured,
-             or "error: <message>" if sending failed.
-    """
-    try:
-        from metrics_utility.library.storage.segment import SEGMENT_AVAILABLE, StorageSegment
-    except ImportError:
-        logger.warning("metrics-utility segment integration not available")
-        return "segment_not_available"
-
-    if not SEGMENT_AVAILABLE or StorageSegment is None:
-        return "segment_not_available"
-
-    try:
-        import json
-
-        from django.conf import settings
-
-        # Get Segment write key from settings
-        write_key = getattr(settings, "SEGMENT_WRITE_KEY", None)
-        if not write_key:
-            logger.warning("SEGMENT_WRITE_KEY not configured in settings")
-            return "segment_not_available"
-
-        # Calculate data size for logging
-        data_size = len(json.dumps(segment_data).encode("utf-8"))
-
-        log_task_execution(
-            "segment_send",
-            "processing",
-            f"Sending data to Segment.com using StorageSegment (Size: {data_size} bytes)",
-        )
-
-        # Initialize StorageSegment with configuration
-        storage = StorageSegment(
-            write_key=write_key,
-            user_id=user_id,
-            debug=getattr(settings, "DEBUG", False),
-            use_bulk=data_size > 24 * 1024,  # Use bulk mode for large payloads
-        )
-
-        # Send data using StorageSegment.put()
-        artifact_name = f"metrics_collection_{user_id}"
-        chunks = storage.put(artifact_name=artifact_name, dict=segment_data, event_name=event_name)
-
-        # Log success with chunk information
-        chunk_count = len(chunks) if chunks else 1
-        logger.info(f"Successfully sent metrics to Segment.com (Size: {data_size} bytes, Chunks: {chunk_count})")
-        return "success"
-
-    except Exception as e:
-        logger.error(f"Error sending data to Segment.com: {str(e)}")
-        return f"error: {str(e)}"
-
-
 def _serialize_args(kwargs):
     if not kwargs:
         return {}
